Added process_cvx_hull to piece class
Will find the convex hull of the piece, store indices in piece,
and find all points on the piece that, within a convex hull defect,
are the furthest away from the convex hull (as indices).

Also made changes to the draw_curve function, now can pass in
the defect indices, will be drawn as a circle with a slightly
smaller radius (colors are greyscale, so couldn't do red).